requirements.txt
pycaret
dotenv
psycopg2-binary
